+++ contracts/escrow/src/core/dispute.rs
@@ -182,7 +182,7 @@ impl DisputeManager {
 
     pub fn dispute_milestone(
         e: &Env,
-        milestone_index: i128,
+        milestone_index: u32,
         signer: Address,
     ) -> Result<Escrow, ContractError> {
         signer.require_auth();
